% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indexGroups.R
\name{indexGroups}
\alias{indexGroups}
\title{Group trimmed sequences into aligning groups}
\usage{
indexGroups(
  blast,
  fasta,
  mode = "any",
  aln.size,
  peptide.once.per.group = TRUE
)
}
\arguments{
\item{blast}{BLAST alignment table to process.}

\item{fasta}{Epitope sequence AAStringset object to process.}

\item{mode}{Grouping method. "any" creates a smaller number of groups, with
individual groups tending to have more members, such that in each group,
each member must align with at least 1 "any" of the other members. "all"
creates a larger number of groups, with individual groups tending to have
fewer members, such that in each group, each member must align with "all"
other members.}

\item{aln.size}{Minimum length of alignment to consider from BLASTp alignments of 'data'.}

\item{peptide.once.per.group}{Logical defining whether or not an epitope is constrained to only include one interval per contributing peptide. Default value TRUE. Only affects alignment mode = "any".}
}
\description{
Group trimmed sequences into aligning groups
}
